Move CompilationUnitCollector into event.collectors package
